+++ ajax/agent.php
@@ -30,6 +30,8 @@
  * ---------------------------------------------------------------------
  */
 
+use Glpi\Http\Response;
+
 $AJAX_INCLUDE = 1;
 include ('../inc/includes.php');
 header("Content-Type: application/json; charset=UTF-8");
@@ -41,7 +43,7 @@ Session::checkLoginUser();
 if (isset($_POST['action']) && isset($_POST['id'])) {
    $agent = new Agent();
    if (!$agent->getFromDB($_POST['id'])) {
-      Toolbox::logWarning('Unable to load agent #' . $_POST['id']);
+      Response::sendError(404, 'Unable to load agent #' . $_POST['id']);
       return;
    };
    $answer = [];