+++ front/impactitem.form.php
@@ -30,6 +30,8 @@
  * ---------------------------------------------------------------------
  */
 
+use Glpi\Http\Response;
+
 include ('../inc/includes.php');
 
 $impact_item = new ImpactItem();
@@ -39,8 +41,7 @@ if (isset($_POST["update"])) {
 
    // Can't update, id is missing
    if ($id === 0) {
-      Toolbox::logWarning("Can't update the target impact item, id is missing");
-      Html::back();
+      Response::sendError(400, "Can't update the target impact item, id is missing", Response::CONTENT_TYPE_TEXT_HTML);
    }
 
    // Load item and check rights